+++ src/Rest/Tools.php
@@ -1,0 +1,36 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Dbp\Relay\CoreBundle\Rest;
+
+use Kekos\MultipartFormDataParser\Parser;
+use Nyholm\Psr7\Factory\Psr17Factory;
+use Symfony\Bridge\PsrHttpMessage\Factory\HttpFoundationFactory;
+use Symfony\Bridge\PsrHttpMessage\Factory\PsrHttpFactory;
+use Symfony\Component\HttpFoundation\Request;
+
+class Tools
+{
+    /**
+     * Makes request parameters available for a multipart/form-data PATCH request.
+     */
+    public static function fixMultipartPatchRequest(Request $request): Request
+    {
+        $uploaded_file_factory = new Psr17Factory();
+        $stream_factory = $uploaded_file_factory;
+
+        $bridgeFactory = new Psr17Factory();
+        $psrHttpFactory = new PsrHttpFactory($bridgeFactory, $bridgeFactory, $bridgeFactory, $bridgeFactory);
+
+        // convert to psr7 to decorate request with multipart formdata
+        $psrRequest = $psrHttpFactory->createRequest($request);
+        $parser = Parser::createFromRequest($psrRequest, $uploaded_file_factory, $stream_factory);
+        $psrRequest = $parser->decorateRequest($psrRequest);
+
+        // convert back to symfony request to get a symfony uploadedFile instead of a PSR-7 uploadedFile
+        $httpFoundationFactory = new HttpFoundationFactory();
+
+        return $httpFoundationFactory->createRequest($psrRequest);
+    }
+}